hr_chatbot_production/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
mysql-connector-python
sentence-transformers
faiss-cpu
transformers
torch
